requirements.txt
alembic==1.10.4
blinker==1.6.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
pluggy==1.0.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
webargs==8.2.0
Werkzeug==2.3.3
